+++ Assets/Scripts/IDamageable.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    public void TakeDamage(int amount);
+    public void TakeDamage(float amount);
 
 }
